Fix the command to execute installation
When you try to install the library it send a error when it try to
install all dependency and run the command to validate all is ok.

With these changes the error is repaired, the only thing necessary was
to add the libraries to run the tests in the main dependencies of the
library.